+++ tests/test_file.py
@@ -1,0 +1,76 @@
+import pytest
+import os
+
+
+from exdir.core import File
+from exdir.core import DATASET_TYPENAME, FILE_TYPENAME
+from exdir.core import _create_object_directory, _is_valid_object_directory
+
+from conftest import remove
+
+
+def test_file_init(setup_teardown_folder):
+    no_exdir = os.path.join(pytest.TESTPATH, "no_exdir")
+
+    f = File(no_exdir, mode="w")
+    f.close()
+    assert(_is_valid_object_directory(no_exdir + ".exdir"))
+    remove(pytest.TESTFILE)
+
+    f = File(pytest.TESTFILE, mode="w")
+    f.close()
+    assert(_is_valid_object_directory(pytest.TESTFILE))
+    remove(pytest.TESTFILE)
+
+    f = File(pytest.TESTFILE, mode="a")
+    f.close()
+    assert(_is_valid_object_directory(pytest.TESTFILE))
+    remove(pytest.TESTFILE)
+
+    f = File(pytest.TESTFILE, mode="a")
+    f.close()
+    assert(_is_valid_object_directory(pytest.TESTFILE))
+    remove(pytest.TESTFILE)
+
+    os.makedirs(pytest.TESTFILE)
+    with pytest.raises(FileExistsError):
+        f = File(pytest.TESTFILE, mode="w")
+
+    remove(pytest.TESTFILE)
+
+    _create_object_directory(pytest.TESTFILE, DATASET_TYPENAME)
+    with pytest.raises(FileExistsError):
+        f = File(pytest.TESTFILE, mode="w")
+
+    remove(pytest.TESTFILE)
+
+    with pytest.raises(IOError):
+        f = File(pytest.TESTFILE, mode="r")
+    with pytest.raises(IOError):
+        f = File(pytest.TESTFILE, mode="r+")
+
+
+    _create_object_directory(pytest.TESTFILE, FILE_TYPENAME)
+
+    with pytest.raises(FileExistsError):
+        f = File(pytest.TESTFILE, mode="w")
+
+    remove(pytest.TESTFILE)
+
+    _create_object_directory(pytest.TESTFILE, FILE_TYPENAME)
+    f = File(pytest.TESTFILE, mode="w", allow_remove=True)
+    remove(pytest.TESTFILE)
+
+    _create_object_directory(pytest.TESTFILE, FILE_TYPENAME)
+
+    with pytest.raises(IOError):
+        f = File(pytest.TESTFILE, mode="w-")
+
+    with pytest.raises(IOError):
+        f = File(pytest.TESTFILE, mode="x")
+
+
+
+def test_file_close(setup_teardown_folder):
+    f = File(pytest.TESTFILE, mode="w")
+    f.close()
